Extract Popup data to separate repository for better code organization:
