+++ app/src/main/java/com/cornellappdev/android/eatery/data/repositories/UserPreferencesRepository.kt
@@ -1,7 +1,6 @@
 package com.cornellappdev.android.eatery.data.repositories
 
 import androidx.datastore.core.DataStore
-import com.cornellappdev.android.eatery.Date
 import com.cornellappdev.android.eatery.UserPreferences
 import com.cornellappdev.android.eatery.util.Constants.PASSWORD_ALIAS
 import com.cornellappdev.android.eatery.util.encryptData
@@ -14,7 +13,6 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import java.time.LocalDate
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -36,50 +34,6 @@ class UserPreferencesRepository @Inject constructor(
     val recentSearchesFlow: StateFlow<List<Int>> = userPreferencesFlow.map { prefs ->
         prefs.recentSearchesList
     }.stateIn(CoroutineScope(Dispatchers.IO), SharingStarted.Eagerly, listOf())
-
-    val lastShowedRatingPopupFlow = userPreferencesFlow.map {
-        with(it.lastShowedRatingPopup) {
-            // Default value should be min local date
-            if (year == 0) LocalDate.MIN else
-                LocalDate.now().withYear(year).withDayOfMonth(day)
-        }
-    }.stateIn(CoroutineScope(Dispatchers.IO), SharingStarted.Eagerly, LocalDate.MIN)
-
-    val minDaysBetweenRatingShow =
-        userPreferencesFlow.map { it.minDaysBetweenRatingShow }
-            .stateIn(CoroutineScope(Dispatchers.IO), SharingStarted.Eagerly, 7)
-
-    init {
-        CoroutineScope(Dispatchers.IO).launch {
-            userPreferencesStore.updateData {
-                /**
-                 * This is a way of manually setting the default value for minDaysBetweenRatingShow
-                 * so we can multiply it for an exponential fall off.
-                 * Currently it doesn't seem like proto files support custom default values so this
-                 * is the only way to do it really.
-                 * See https://protobuf.dev/programming-guides/proto3/#default for more details
-                 */
-                with(it.toBuilder()) {
-                    if (it.minDaysBetweenRatingShow == 0) {
-                        setMinDaysBetweenRatingShow(7).build()
-                    } else {
-                        build()
-                    }
-                }
-            }
-        }
-    }
-
-    suspend fun onShownRating() {
-        userPreferencesStore.updateData { prefs ->
-            with(prefs.toBuilder()) {
-                setMinDaysBetweenRatingShow((prefs.minDaysBetweenRatingShow * 1.5).toInt())
-                setLastShowedRatingPopup(
-                    Date.newBuilder().setYear(2024).setMonth(10).setDay(12)
-                ).build()
-            }
-        }
-    }
 
     suspend fun setHasOnboarded(hasOnboarded: Boolean) {
         userPreferencesStore.updateData { currentPreferences ->
